Modified: Data type in methods to accept float values and return the same

+++ main/java/com/capgemini/genericstestmaximum/FindMaximum.java
@@ -14,7 +14,7 @@ public class FindMaximum
 	 * @param z
 	 * @return maximum
 	 */
-	public static Integer findMaximum(Integer x,Integer y,Integer z) {
+	public static Float findMaximum(Float x,Float y,Float z) {
 		int n1 = x.compareTo(y);
 		int n2 = y.compareTo(z);
 		int n3 = x.compareTo(z);
@@ -29,18 +29,18 @@ public class FindMaximum
 		else if(n1==1&&n3==-1)
 			return z;
 		else 
-			return 0;
+			return 0.0f;
 	}
     public static void main( String[] args )
     {
     	Scanner sc = new Scanner(System.in);
         System.out.println("Enter 3 integers to find maximum");
         System.out.println("Enter first integer");
-        Integer x = Integer.parseInt(sc.nextLine());
+        Float x = Float.parseFloat(sc.nextLine());
         System.out.println("Enter second integer");
-        Integer y = Integer.parseInt(sc.nextLine());
+        Float y = Float.parseFloat(sc.nextLine());
         System.out.println("Enter third integer");
-        Integer z = Integer.parseInt(sc.nextLine());
+        Float z = Float.parseFloat(sc.nextLine());
         System.out.println(findMaximum(x,y,z));
     }
 }
